Cleaned up the code.
Converted a few more comments to english.
Removed the dependecy on glos.h
Now the calculations are done with floats.

+++ pi.h
@@ -22,6 +22,6 @@
 
 const float PI = 3.1415927f;
 
-inline double UTIL_TO_RADIANS(double angle) { return angle * PI / 180; }
+inline float UTIL_TO_RADIANS(float angle) { return angle * PI / 180; }
 
 #endif /* PI_H_ */
